Remove insert latency burstiness
Insertion latency sometimes shows large spikes due to two defects
in NuDB's write throttling code.

1. If a burst of inserts occurs immediately after a batch write
finishes, they may temporarily exceed the proven sustained write
rate. This is common and doesn't require any throttling.

2. When a small batch write completes, the execution time may be
dominated by overhead, resulting in an artificially low
measurement of the sustained store rate. We don't want to replace
our measurement from a large batch with that one.

This change adds a default "burst size" of 4MB. For write batches
below this size, we do not update our measurement of the
sustained store writes. And until the size of the new write batch
reaches this size, we do not throttle writes.

The burst size can be set with a "set_burst" function. Setting
a burst size of zero restores the original behavior. Tuning this
should not be necessary.